backend/requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
SQLAlchemy==2.0.30
pydantic==2.7.4
pydantic-settings==2.3.3
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
requests==2.31.0
httpx==0.27.0
PyYAML==6.0.1

